requirement.txt
aiosqlite==0.16.1
asyncpg==0.22.0
bcrypt==3.2.0
cffi==1.14.5
Deprecated==1.2.11
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
fastapi-pagination==0.5.2
fastapi-users==5.1.0
idna==3.1
iso8601==0.1.14
makefun==1.9.5
passlib==1.7.4
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.1
PyPika==0.44.1
python-multipart==0.0.5
pytz==2020.5
six==1.15.0
starlette==0.13.6
tortoise-orm==0.16.21
typing-extensions==3.7.4.3
wrapt==1.12.1
